Fix biased bananajack shuffle
Use a Fisher-Yates shuffle instead of a (biased) 52^52 shuffle.
The biased shuffle raises the expected value of the
starting hands slightly, particularly for the dealer.  As an
example, the dealer's first card is K♥ with 2.6% probability
and 2♠ with only 1.43%, while the unbiased probability of any
card with the correct algorithm is 1.92%.

See also:
https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle#Implementation_errors